QnA/requirements.txt
# For MLFlow evaluation
mlflow==3.3.2
pandas==2.3.2
evaluate==0.4.5
textstat==0.7.8
tiktoken==0.11.0
torch==2.8.0
tokenizers==0.21.4
transformers==4.55.4
sentence-transformers==4.1.0
openpyxl==3.1.5
python-dotenv==1.1.1
openai==1.109.1
